+++ native/ios/BulletANE/BulletANE/CylinderShapeAPI.cpp
@@ -24,7 +24,7 @@ extern "C" FREObject createCylinderShape(FREContext ctx, void *funcData, uint32_
     btCollisionShape* shape = new btCylinderShape(halfExtents);
     
     FREObject ptr;
-    FRENewObjectFromUint32((uint32_t)shape, &ptr);
+    FRENewObjectFromUint32((uint32_t)(size_t)shape, &ptr);
     return ptr;
 }
